- Updated code, README and remove old cifar10
