+++ examples/contrib/cifar10/main.py
@@ -1,37 +1,126 @@
-import argparse
 from pathlib import Path
+from datetime import datetime
+
+import fire
 
 import torch
 import torch.nn as nn
-import torch.distributed as dist
-
+import torch.optim as optim
 
 import ignite
+import ignite.distributed as idist
 from ignite.engine import Events, Engine, create_supervised_evaluator
-from ignite.engine.deterministic import DeterministicEngine
 from ignite.metrics import Accuracy, Loss
-from ignite.handlers import Checkpoint, global_step_from_engine
-from ignite.utils import convert_tensor, manual_seed
+from ignite.handlers import Checkpoint
+from ignite.utils import manual_seed
 
 from ignite.contrib.engines import common
-from ignite.contrib.handlers import TensorboardLogger, ProgressBar
-from ignite.contrib.handlers.tensorboard_logger import OutputHandler, OptimizerParamsHandler, GradsHistHandler
+from ignite.contrib.handlers import ProgressBar
 from ignite.contrib.handlers import PiecewiseLinear
 
 import utils
 
 
-def run(output_path, config):
+def get_dataflow(config):
+    # - Get train/test datasets
+    if idist.get_rank() > 0:
+        # Ensure that only rank 0 download the dataset
+        idist.barrier()
 
-    distributed = dist.is_available() and dist.is_initialized()
-    rank = dist.get_rank() if distributed else 0
+    train_dataset, test_dataset = utils.get_train_test_datasets(config["data_path"])
 
+    if idist.get_rank() == 0:
+        # Ensure that only rank 0 download the dataset
+        idist.barrier()
+
+    # Setup data loader also adapted to distributed config: nccl, gloo, xla-tpu
+    train_loader = idist.auto_dataloader(
+        train_dataset,
+        batch_size=config["batch_size"],
+        num_workers=config["num_workers"],
+        shuffle=True,
+        pin_memory="cuda" in idist.device().type,
+        drop_last=True,
+    )
+
+    test_loader = idist.auto_dataloader(
+        test_dataset,
+        batch_size=2 * config["batch_size"],
+        num_workers=config["num_workers"],
+        shuffle=False,
+        pin_memory="cuda" in idist.device().type,
+    )
+    return train_loader, test_loader
+
+
+def initialize(config):
+    model = utils.get_model(config["model"])
+    # Adapt model for distributed settings if configured
+    model = idist.auto_model(model)
+
+    optimizer = optim.SGD(
+        model.parameters(),
+        lr=config["learning_rate"],
+        momentum=config["momentum"],
+        weight_decay=config["weight_decay"],
+        nesterov=True,
+    )
+    optimizer = idist.auto_optim(optimizer)
+    criterion = nn.CrossEntropyLoss().to(idist.device())
+    return model, optimizer, criterion
+
+
+def print_metrics(epoch, tag, metrics):
+    if idist.get_rank() == 0:
+        print(
+            "\nEpoch {} - {} metrics:\n {}".format(
+                epoch, tag, "\n".join(["\t{}: {}".format(k, v) for k, v in metrics.items()])
+            )
+        )
+
+
+def training(local_rank, config):
+
+    rank = idist.get_rank()
     manual_seed(config["seed"] + rank)
+    device = idist.device()
+
+    if rank == 0:
+        print("Train {} on CIFAR10".format(config["model"]))
+        print("- PyTorch version: {}".format(torch.__version__))
+        print("- Ignite version: {}".format(ignite.__version__))
+
+        print("\n")
+        print("Configuration:")
+        for key, value in config.items():
+            print("\t{}: {}".format(key, value))
+        print("\n")
+
+        if idist.get_world_size() > 1:
+            print("\nDistributed setting:")
+            print("\tbackend: {}".format(idist.backend()))
+            print("\tworld size: {}".format(idist.get_world_size()))
+            print("\n")
+
+    output_path = config["output_path"]
+    if rank == 0:
+        if config["stop_iteration"] is None:
+            now = datetime.now().strftime("%Y%m%d-%H%M%S")
+        else:
+            now = "stop-on-{}".format(config["stop_iteration"])
+
+        folder_name = "{}_backend-{}-{}_{}".format(config["model"], idist.backend(), idist.get_world_size(), now)
+        output_path = Path(output_path) / folder_name
+        if not output_path.exists():
+            output_path.mkdir(parents=True)
+        config["output_path"] = output_path.as_posix()
+        print("Output path: {}".format(config["output_path"]))
 
     # Setup dataflow, model, optimizer, criterion
-    train_loader, test_loader = utils.get_dataflow(config, distributed)
-    model, optimizer = utils.get_model_optimizer(config, distributed)
-    criterion = nn.CrossEntropyLoss().to(utils.device)
+    train_loader, test_loader = get_dataflow(config)
+
+    # Setup model, optimizer
+    model, optimizer, criterion = initialize(config)
 
     le = len(train_loader)
     milestones_values = [
@@ -52,8 +141,11 @@ def run(output_path, config):
 
     def train_step(engine, batch):
 
-        x = convert_tensor(batch[0], device=utils.device, non_blocking=True)
-        y = convert_tensor(batch[1], device=utils.device, non_blocking=True)
+        x, y = batch[0], batch[1]
+
+        if x.device != device:
+            x = x.to(device, non_blocking=True)
+            y = y.to(device, non_blocking=True)
 
         model.train()
         # Supervised part
@@ -64,112 +156,96 @@ def run(output_path, config):
         loss.backward()
         optimizer.step()
 
+        # This can be helpful for XLA to avoid performance slow down if fetch loss.item() every iteration
+        if config["log_every_iters"] > 0 and (engine.state.iteration - 1) % config["log_every_iters"] == 0:
+            batch_loss = loss.item()
+            engine.state.saved_batch_loss = batch_loss
+        else:
+            batch_loss = engine.state.saved_batch_loss
+
         return {
-            "batch loss": loss.item(),
+            "batch loss": batch_loss,
         }
 
-    if config["deterministic"] and rank == 0:
-        print("Setup deterministic trainer")
-    trainer = Engine(train_step) if not config["deterministic"] else DeterministicEngine(train_step)
-    train_sampler = train_loader.sampler if distributed else None
+    trainer = Engine(train_step)
+    trainer.state.saved_batch_loss = -1.0
+    trainer.state_dict_user_keys.append("saved_batch_loss")
+
     to_save = {"trainer": trainer, "model": model, "optimizer": optimizer, "lr_scheduler": lr_scheduler}
     metric_names = [
         "batch loss",
     ]
+    output_path = config["output_path"]
     common.setup_common_training_handlers(
         trainer,
-        train_sampler=train_sampler,
+        train_sampler=train_loader.sampler,
         to_save=to_save,
         save_every_iters=config["checkpoint_every"],
         output_path=output_path,
         lr_scheduler=lr_scheduler,
-        output_names=metric_names,
-        with_pbar_on_iters=config["display_iters"],
-        log_every_iters=10,
+        output_names=metric_names if config["log_every_iters"] > 0 else None,
+        with_pbar_on_iters=config["log_every_iters"] > 0,
+        log_every_iters=config["log_every_iters"],
+        clear_cuda_cache=False,
     )
-
-    if rank == 0:
-        # Setup Tensorboard logger - wrapper on SummaryWriter
-        tb_logger = TensorboardLogger(log_dir=output_path)
-        # Attach logger to the trainer and log trainer's metrics (stored in trainer.state.metrics) every iteration
-        tb_logger.attach(
-            trainer,
-            log_handler=OutputHandler(tag="train", metric_names=metric_names),
-            event_name=Events.ITERATION_COMPLETED,
-        )
-        # log optimizer's parameters: "lr" every iteration
-        tb_logger.attach(
-            trainer, log_handler=OptimizerParamsHandler(optimizer, param_name="lr"), event_name=Events.ITERATION_STARTED
-        )
 
     # Let's now setup evaluator engine to perform model's validation and compute metrics
     metrics = {
-        "accuracy": Accuracy(device=utils.device if distributed else None),
-        "loss": Loss(criterion, device=utils.device if distributed else None),
+        "accuracy": Accuracy(),
+        "loss": Loss(criterion),
     }
 
     # We define two evaluators as they wont have exactly similar roles:
     # - `evaluator` will save the best model based on validation score
-    evaluator = create_supervised_evaluator(model, metrics=metrics, device=utils.device, non_blocking=True)
-    train_evaluator = create_supervised_evaluator(model, metrics=metrics, device=utils.device, non_blocking=True)
+    evaluator = create_supervised_evaluator(model, metrics=metrics, device=device, non_blocking=True)
+    train_evaluator = create_supervised_evaluator(model, metrics=metrics, device=device, non_blocking=True)
 
     def run_validation(engine):
-        train_evaluator.run(train_loader)
-        evaluator.run(test_loader)
+        epoch = trainer.state.epoch
+        state = train_evaluator.run(train_loader)
+        print_metrics(epoch, "Train", state.metrics)
+        state = evaluator.run(test_loader)
+        print_metrics(epoch, "Test", state.metrics)
 
-    trainer.add_event_handler(Events.EPOCH_STARTED(every=config["validate_every"]), run_validation)
-    trainer.add_event_handler(Events.COMPLETED, run_validation)
+    trainer.add_event_handler(Events.EPOCH_COMPLETED(every=config["validate_every"]) | Events.COMPLETED, run_validation)
 
     if rank == 0:
         # Setup progress bar on evaluation engines
-        if config["display_iters"]:
+        if config["log_every_iters"] > 0:
             ProgressBar(persist=False, desc="Train evaluation").attach(train_evaluator)
             ProgressBar(persist=False, desc="Test evaluation").attach(evaluator)
 
-        # Let's log metrics of `train_evaluator` stored in `train_evaluator.state.metrics` when validation run is done
-        tb_logger.attach(
-            train_evaluator,
-            log_handler=OutputHandler(
-                tag="train", metric_names="all", global_step_transform=global_step_from_engine(trainer)
-            ),
-            event_name=Events.COMPLETED,
+        # Setup TensorBoard logging on trainer and evaluators. Logged values are:
+        #  - Training metrics, e.g. running average loss values
+        #  - Learning rate
+        #  - Evaluation train/test metrics
+        evaluators = {"training": train_evaluator, "test": evaluator}
+        tb_logger = common.setup_tb_logging(output_path, trainer, optimizer, evaluators=evaluators)
+
+        trains_logger = common.setup_trains_logging(
+            trainer, optimizer, evaluators=evaluators, project_name="cifar10-ignite", task_name=Path(output_path).stem
         )
 
-        # Let's log metrics of `evaluator` stored in `evaluator.state.metrics` when validation run is done
-        tb_logger.attach(
-            evaluator,
-            log_handler=OutputHandler(
-                tag="test", metric_names="all", global_step_transform=global_step_from_engine(trainer)
-            ),
-            event_name=Events.COMPLETED,
-        )
+    # Store 3 best models by validation accuracy:
+    common.save_best_model_by_val_score(
+        output_path, evaluator, model=model, metric_name="accuracy", n_saved=3, trainer=trainer, tag="test"
+    )
 
-        # Store 3 best models by validation accuracy:
-        common.save_best_model_by_val_score(
-            output_path, evaluator, model=model, metric_name="accuracy", n_saved=3, trainer=trainer, tag="test"
-        )
+    # In order to check training resuming we can stop training on a given iteration
+    if config["stop_iteration"] is not None:
 
-        # Optionally log model gradients
-        if config["log_model_grads_every"] is not None:
-            tb_logger.attach(
-                trainer,
-                log_handler=GradsHistHandler(model, tag=model.__class__.__name__),
-                event_name=Events.ITERATION_COMPLETED(every=config["log_model_grads_every"]),
-            )
-
-    # In order to check training resuming we can emulate a crash
-    if config["crash_iteration"] is not None:
-
-        @trainer.on(Events.ITERATION_STARTED(once=config["crash_iteration"]))
-        def _(engine):
-            raise Exception("STOP at iteration: {}".format(engine.state.iteration))
+        @trainer.on(Events.ITERATION_STARTED(once=config["stop_iteration"]))
+        def _():
+            print("Stop training on {} iteration".format(trainer.state.iteration))
+            trainer.terminate()
 
     resume_from = config["resume_from"]
     if resume_from is not None:
         checkpoint_fp = Path(resume_from)
         assert checkpoint_fp.exists(), "Checkpoint '{}' is not found".format(checkpoint_fp.as_posix())
-        print("Resume from a checkpoint: {}".format(checkpoint_fp.as_posix()))
-        checkpoint = torch.load(checkpoint_fp.as_posix())
+        if rank == 0:
+            print("Resume from a checkpoint: {}".format(checkpoint_fp.as_posix()))
+        checkpoint = torch.load(checkpoint_fp.as_posix(), map_location="cpu")
         Checkpoint.load_objects(to_load=to_save, checkpoint=checkpoint)
 
     try:
@@ -181,87 +257,67 @@ def run(output_path, config):
 
     if rank == 0:
         tb_logger.close()
+        trains_logger.close()
+
+
+def run(
+    seed=543,
+    data_path="/tmp/cifar10",
+    output_path="/tmp/output-cifar10/",
+    model="resnet18",
+    batch_size=512,
+    momentum=0.9,
+    weight_decay=1e-4,
+    num_workers=12,
+    num_epochs=24,
+    learning_rate=0.4,
+    num_warmup_epochs=4,
+    validate_every=3,
+    checkpoint_every=200,
+    backend=None,
+    resume_from=None,
+    log_every_iters=15,
+    num_procs_per_node=None,
+    stop_iteration=None,
+    **spawn_kwargs
+):
+    """Main entry to train an model on CIFAR10 dataset.
+
+    Args:
+        seed (int): random state seed to set. Default, 543.
+        data_path (str): input dataset path. Default, "/tmp/cifar10".
+        output_path (str): output path. Default, "/tmp/output-cifar10".
+        model (str): model name (from torchvision) to setup model to train. Default, "resnet18".
+        batch_size (int): total batch size. Default, 512.
+        momentum (float): optimizer's momentum. Default, 0.9.
+        weight_decay (float): weight decay. Default, 1e-4.
+        num_workers (int): number of workers in the data loader. Default, 5.
+        num_epochs (int): number of epochs to train the model. Default, 24.
+        learning_rate (float): peak of piecewise linear learning rate scheduler. Default, 0.4.
+        num_warmup_epochs (int): number of warm-up epochs before learning rate decay. Default, 4.
+        validate_every (int): run model's validatation every ``validate_every`` epochs. Default, 3.
+        checkpoint_every (int): store training checkpoint every ``checkpoint_every`` iterations. Default, 200.
+        backend (str, optional): backend to use for distributed configuration. Possible values: None, "nccl", "xla-tpu",
+            "gloo" etc. Default, None.
+        num_procs_per_node (int, optional): optional argument to setup number of processes per node. It is useful,
+            when main python process is spawning training as child processes.
+        resume_from (str, optional): path to checkpoint to use to resume the training from. Default, None.
+        log_every_iters (int): argument to log progress every ``log_every_iters`` iterations. It can be 0 to disable it.
+            Default, 15.
+        stop_iteration (int, optional): iteration to stop the training. Can be used to check resume from checkpoint.
+        **spawn_kwargs: Other kwargs to spawn training as child processes.
+
+    """
+    # catch all local parameters
+    config = locals()
+    config.update(config["spawn_kwargs"])
+    del config["spawn_kwargs"]
+
+    spawn_kwargs["num_procs_per_node"] = num_procs_per_node
+
+    with idist.Parallel(backend=backend, **spawn_kwargs) as parallel:
+        parallel.run(training, config)
 
 
 if __name__ == "__main__":
-
-    parser = argparse.ArgumentParser("Training a CNN on CIFAR10 dataset")
-    parser.add_argument(
-        "--params",
-        type=str,
-        help="Override default configuration with parameters: "
-        "data_path=/path/to/dataset;batch_size=64;num_workers=12 ...",
-    )
-    parser.add_argument("--local_rank", type=int, default=0, help="Local process rank in distributed computation")
-
-    args = parser.parse_args()
-
-    assert torch.cuda.is_available()
-    torch.backends.cudnn.benchmark = True
-
-    config = utils.get_default_config()
-    config["local_rank"] = args.local_rank
-
-    # Override config:
-    if args.params is not None:
-        for param in args.params.split(";"):
-            key, value = param.split("=")
-            if "/" not in value:
-                value = eval(value)
-            config[key] = value
-
-    backend = config["dist_backend"]
-    distributed = backend is not None
-
-    if distributed:
-        dist.init_process_group(backend, init_method=config["dist_url"])
-        # let each node print the info
-        if config["local_rank"] == 0:
-            print("\nDistributed setting:")
-            print("\tbackend: {}".format(dist.get_backend()))
-            print("\tworld size: {}".format(dist.get_world_size()))
-            print("\trank: {}".format(dist.get_rank()))
-            print("\n")
-
-    output_path = None
-    # let each node print the info
-    if config["local_rank"] == 0:
-        print("Train {} on CIFAR10".format(config["model"]))
-        print("- PyTorch version: {}".format(torch.__version__))
-        print("- Ignite version: {}".format(ignite.__version__))
-        print("- CUDA version: {}".format(torch.version.cuda))
-
-        print("\n")
-        print("Configuration:")
-        for key, value in config.items():
-            print("\t{}: {}".format(key, value))
-        print("\n")
-
-        # create log directory only by 1 node
-        if (not distributed) or (dist.get_rank() == 0):
-            from datetime import datetime
-
-            now = datetime.now().strftime("%Y%m%d-%H%M%S")
-            gpu_conf = "-single-gpu"
-            if distributed:
-                ngpus_per_node = torch.cuda.device_count()
-                nnodes = dist.get_world_size() // ngpus_per_node
-                gpu_conf = "-distributed-{}nodes-{}gpus".format(nnodes, ngpus_per_node)
-
-            output_path = Path(config["output_path"]) / "{}{}".format(now, gpu_conf)
-            if not output_path.exists():
-                output_path.mkdir(parents=True)
-            output_path = output_path.as_posix()
-            print("Output path: {}".format(output_path))
-
-    try:
-        run(output_path, config)
-    except KeyboardInterrupt:
-        print("Catched KeyboardInterrupt -> exit")
-    except Exception as e:
-        if distributed:
-            dist.destroy_process_group()
-        raise e
-
-    if distributed:
-        dist.destroy_process_group()
+    fire.Fire({"run": run})
