--- examples/contrib/cifar10/plx_configs/xp_training_1n_1gpu.yaml.tmpl
@@ -1,28 +1,0 @@
----
-version: 1
-kind: experiment
-
-tags: [training, 1n1gpu]
-
-# Setup running node:
-environment:
-  resources:
-      gpu:
-        requests: 1
-        limits: 1
-
-# Setup running environment:
-build:
-  image: pytorch/pytorch:latest
-  build_steps:
-
-  # Install ignite and other useful packages
-  - pip install --upgrade git+https://github.com/pytorch/ignite.git &&
-    pip install tqdm tensorboardX
-
-run:
-  cmd:
-  # Setup a path to already downloaded CIFAR10 dataset
-  # If not the case, just set a non-existing folder to where to download the dataset
-  - export DATA_PATH="/tmp/cifar10"
-  - python main.py --params="data_path=$DATA_PATH;output_path=$POLYAXON_RUN_OUTPUTS_PATH;display_iters=False"
