--- examples/contrib/cifar10/plx_configs/xp_training_2n_4gpus.yaml.tmpl
@@ -1,49 +1,0 @@
----
-version: 1
-kind: experiment
-framework: pytorch
-
-tags: [training, 2n4gpus]
-
-# Setup running node:
-environment:
-  resources:
-      gpu:
-        requests: 2
-        limits: 2
-
-  replicas: 
-    n_workers: 1
-    default_worker:
-      resources:
-        gpu:
-          requests: 2
-          limits: 2
-  
-
-# Setup running environment:
-build:
-  image: pytorch/pytorch:latest
-  build_steps:
-
-  # Install ignite and other useful packages
-  - pip install --upgrade git+https://github.com/pytorch/ignite.git &&
-    pip install tqdm tensorboardX
-
-run:
-  cmd:
-  - sleep 5
-
-  # Setup a path to already downloaded CIFAR10 dataset
-  # If not the case, just set a non-existing folder to where to download the dataset
-
-  - export DATA_PATH="/tmp/cifar10"
-  - echo "MASTER_ADDR=$MASTER_ADDR"
-  - echo "MASTER_PORT=$MASTER_PORT"
-  - echo "nnodes=$WORLD_SIZE"
-  - echo "node_rank=$RANK"
-
-  - python -m torch.distributed.launch --nproc_per_node=2
-    --nnodes=$WORLD_SIZE --node_rank=$RANK --master_addr="$MASTER_ADDR" --master_port=$MASTER_PORT
-    main.py --params="batch_size=512;dist_backend='nccl';data_path=$DATA_PATH;output_path=$POLYAXON_RUN_OUTPUTS_PATH;display_iters=False"
-
