--- examples/contrib/new-cifar10/.gitignore
@@ -1,5 +1,0 @@
-output
-cifar10
-.polyaxonignore
-.polyaxon
-plx_configs/*.yaml
